requirements.txt
pandas
numpy
matplotlib
pyyaml
yfinance
fredapi
streamlit
requests
sec-edgar-downloader
python-dotenv
